DDCNLS-7107 | WIP
